+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.0.0 / 2019-11-07
+==================
+
+* InitService and closeService are now init and close respectively
+* MessageConcurrency option renamed to concurrency
+
 1.3.0 / 2019-10-12
 ==================